requirements.txt
ccxt
pandas
pyyaml